server/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
idna==3.8
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
channels==4.0.0
channels-redis==4.1.0
django-redis==5.4.0
psycopg2-binary==2.9.9
paho-mqtt==1.6.1
daphne==4.0.0
